send url and get json reponse
