+++ Cargo.toml
@@ -7,6 +7,14 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+rocket = "0.4.7"
 headless_chrome = "0.9.0"
 failure = "0.1.8"
 url = "2.2.1"
+serde = "1.0.125"
+serde_json = "1.0.64"
+
+[dependencies.rocket_contrib]
+version = "0.4.7"
+default-features = false
+features = ["json"]
